Move grpc methodHandler to separate file to control over logging separately
